addressing the sep files comments pt 2
